Dynamically set the native exception signal handler to SIG_DFL
  The native exception signal handler will be called when a process
exception occurs. Because of the long flow of exception handling,
if a signal 9 is received in the middle, the whole process will be
interrupted, and there may be no tombstone file generated.

   OEM wants to control whether signal handler is dynamically set to
SIG_DFL through environment variable in the debuggable load. OEM can
export an environment variable in RC file to control the value of this
environment variable is true or false. When the native exception signal
handler is SIG_DFL, OEM can obtain the process coredump directly by 
configuring /proc/sys/ kernel/core_pattern in the debuggable load.

Change-Id: If894b4582a1820b64bdae819cec593b7710cb6e3