Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> From the other thread: if this is required only for userdata, then maybe better to only enable this flag for /data?

Yes, and I've moved to code? :)

> Also, what about other RW partitions?

I'm not sure if there will be use cases for other RW partitions.  We could enable for all of them as there should be no overhead at all.  Eric can confirm.  Only f2fs and ext4 are supported at the moment, so we'll need to check that.  This seems to be better as we won't need to worry about this again.  WDYT?

Patch-set: 1
Reviewer: Gerrit User 1042276 <1042276@85c56323-6fa9-3386-8a01-6480fb634889>