Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> > Patch Set 1:
> > 
> > From the other thread: if this is required only for userdata, then maybe better to only enable this flag for /data?
> 
> Yes, and I've moved to code? :)

Ah got it.  That's not likely the best place to put it; that function is used to filter out the /data fstab entry once we've already mounted it and know where it is.  But that isn't the only way the /data is referenced; it's directly mounted in fs_mgr_mount_all() without the use of that function for example.  You'll want to keep these lines in their original location when the fstab is being parsed, but check the mount_point for /data.

> 
> > Also, what about other RW partitions?
> 
> I'm not sure if there will be use cases for other RW partitions.  We could enable for all of them as there should be no overhead at all.  Eric can confirm.  Only f2fs and ext4 are supported at the moment, so we'll need to check that.  This seems to be better as we won't need to worry about this again.  WDYT?

If you're okay not enabling it on the other RW partitions, then we're okay with this.  It'll just be hard if you do want to, since some of the RW partitions are mounted in first stage mount, so before properties are online.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>