WIP: Replace the hashtable of entries with a std::map
As we increase the range of the offsets to 8 bytes, the size of the struct
ZipStringOffset and std::string_view will be the same. Both will have 16
bytes after alignment. Plus, the hashtable needs a continuous chunk of
memory, and we are overallocating at least by 25%. On the otherhand, a
std::map<std::string_view, uint64_t> uses 24 bytes + pointers overhead.

Test: build
Change-Id: I74d95f53207cc8ca871b955e2a15c184d5497833