Update patch set 5
Patch Set 5:

i guess we know from the file size whether we'll ever need a large _offset_ (even if the file is zip64 format, if it's less than 4GiB we know no offset is too large for a uint32_t), but we can't know until we've been through all the entries whether we have any long names. (but that does suggest an alternative to the interface where we choose one of two implementations up front --- instead we could always have both a "small" map and a "large" map, and any name whose offset & length fit in the small map go in the small map, and vice versa. on lookup, we'd know which map to check for long names, but for short names we'd have to check both maps, which is a bit unfortunate. but if we check the short map first, this should mean you're only paying for what you use --- 99% of callers will only fill out their "small" map, and will only ever check their "small" map when doing lookups. only those 1% of callers who actually _need_ long offsets or names end up constructing [or checking] a "large" map.)

Patch-set: 5