Update patch set 2
Patch Set 2: Code-Review+2

You'll need enh@ for owners on the rest, but lgtm assuming that this is a mechanical change and doesn't change the policy.

It may be worth a comment in libcutils that there isn't actually consensus that it should be APEX available.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2