Update patch set 1
Patch Set 1:

> Patch Set 1: Code-Review-1
 > 
 > Please let's not do this.  Please see https://android-review.googlesource.com/c/platform/system/core/+/959678 , which was a similar attempt to do this.
 > 
 > 1) Skipping honoring sehash is just working around other bugs. The label of the file should never be in an incorrect state, and if it is, it's up to the code to fix the label. I don't want to encourage this hack around other bugs in the system.
 > 2) The hashing is really important to maintain boot time performance.

I understand, what if an OEM relabel a file in /persist and the new context isn't propagated properly how it happens in our case? This seems a bug to me
All we did is take stock /persist, relabel a few files on our rules, run restorecon on it and had this.

Patch-set: 1
Reviewer: Gerrit User 1081223 <1081223@85c56323-6fa9-3386-8a01-6480fb634889>