Update patch set 4
Patch Set 4:

Just to confirm - this is a different mechanism than the "firmware usermode helper", right?  We do use that mechanism; ueventd has logic to serve firmwares to the kernel if they are requested through uevents.  

This looks to be a different mechanism that attempts to call modprobe directly?  I agree; I don't believe that we use that anywhere on Android.

Patch-set: 4
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>