Update patch set 1
Patch Set 1: Code-Review+2

I would kind of prefer that the implementation of the pc adjustment be a static function in each Regs object. But the code is small except for the arm one, so I don't think it's worth it.

Patch-set: 1
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Code-Review=+2