Update patch set 1
Patch Set 1:

Do these need to be read-only?  I've been trying to discourage new vendor readable ro. properties, since if we ever have a use case for them to be writeable, it would be API breaking, since the "ro." is part of the name which is part of the API.

SEPolicy should be enough to lockdown the writing, since it limits it to only specific domains.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>