Add a timeout parameter to write, chmod and chown in builtin cmd
the new command format is as following:
write <file_path> <value> [timeout]
chmod <permission> <file_path> [timeout]
chown <uid> [gid] <file_path> [timeout]
The timeout is optional, if it's not used, the command is consistent
with original, otherwise after the command fails, it waits util the
timeout or the file_path is created and retries.

if you don't known when the file_path is created, you can try to add
timeout to the command tail.

Test: add 2 at chmod line tail, after reboot to check kernel boot log.
Change-Id: Ia5b334224cccfee55111a38ba001567f002e8347