Update patch set 1
Patch Set 1:

Is this because we're running the test as multi-arch but we're only building libfs_mgr for the first arch?  In the past we've just dropped running it on the second arch, since libdm is not going to be used in any second arch processes.

If no, do you know what the net size impact of the device is?  Do we still have a shared libfs_mgr but now we're duplicating it in libdm as well?  We'd be duplicating it in process as well, since init included libfs_mgr as a shared library.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>