Add a version field to android_log_header_t
The protocol to write messages to logd has been defacto stable for a
while, and we are now explicitly making that guarantee.

To give flexibility to add to this protocol in the future, we are
adding a version field now, taking from the top three unused bits of
the id field.  This allows for 8 versions and 32 ids, both should be
plenty.  Note that the lack of a size field for the header means that
all new log information will have to be appended to the end of the
message.

Bug: 152466437
Test: liblog-unit-tests, logd-unit-tests
Test: bump version to '1', logs work, above tests pass
Change-Id: I748285d3d1b3c25380e99ae44f7534df744dc669