Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> don't you want some kind of "if version != 0, ignore"?

In logd?  I don't think so.  I explicitly want to accept messages where the version is higher, in case we have an APEX that has knowledge of the new protocol and is running on an older platform with an older logd.

I don't think we'll be able to make breaking changes to the existing protocol.  If we do, we'll likely be better to have a new socket in logd where we can start from scratch and have a header length, etc.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0