base: make libbase logging fork safe
libbase uses a mutex to ensure multiline outputs do not get
interleaved when logging from multiple threads.  This causes problems
if logging after fork() is called, since a different thread may have
locked the mutex, making it impossible for the child to unlock.

This solves this issue by adding a pthread_atfork() handler in
InitLogging(), that 'resets' the lock in the child by reinitializing
the underlying pthread_mutex_t.  Note that this is dangerous in the
following conditions:
1) The thread that calls fork() already owns the mutex, since when the
   thread eventually calls unlock(), the mutex won't have been locked
   anymore.  This can only happen if LogLine itself calls fork(),
   which is not a reasonable use case.
2) APEX's running on pre-R, if they concurrently call SetLogger() and
   fork(), again not a reasonable use case.

This doesn't fix the general issue of multi-threaded APEX's and
fork().  For example, if an APEX uses a different instance of libbase
than the one that called InitLogging(), it will have its own
LoggingLock() that will not be reset in the pthread_atfork() handler.

Bug: 65062446
Bug: 153824050
Test: libbase/liblog unit tests
Test: new unit test for this specific scenario
Change-Id: I44e80b6d65eb1446c841ebcf2f7eaae67096af34