base: make libbase logging fork safe
libbase uses a mutex to ensure multiline outputs do not get
interleaved when logging from multiple threads.  This causes problems
if logging after fork() is called, since a different thread may have
locked the mutex, making it impossible for the child to unlock.

This solves this issue by adding a pthread_atfork() handler in
InitLogging(), that acquires the mutex in the parent and then releases
it in both the child and parent after forking.

Bug: 65062446
Bug: 153824050
Test: libbase/liblog unit tests
Test: new unit test for this specific scenario
Change-Id: I44e80b6d65eb1446c841ebcf2f7eaae67096af34