Update patch set 4
Patch Set 4: Presubmit-Ready+1

I modified this slightly to better isolate the mte code from the rest of the code so it's really easy to build without it.

I don't have a way to test the mte part (I did make sure it compiles properly though).

It would be nice if the bionic mte_supported function in mte,h was not behind a define, but returned false on non-aarch64 systems. This would allow me to remove a couple of defines in the MemoryMteTest.cpp file.

Patch-set: 4
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1