base: APEX and fork safe line splitting
This CL solves two problems:
1) Using LoggingLock to prevent interleaving is not sufficient for
   APEX processes that have multiple instances of libbase, since each
   instance will have its own LoggingLock, and thus there is no
   protection between logs written between instances.
2) The usage of LoggingLock makes libbase logging not safe to use
   after fork().

To solve the first issue, the line splitting code is moved to a lambda
that wraps the provided logger functions.  This lambda contains its
own lock and since all logs will be handled through a single logger
function, it ensure that no logs will be interleaved regardless of
which instance of libbase writes them.

To solve the second issue, a new SetLoggerNoInterleaveLock() is added
that uses a lambda for the line splitting but does not use a lock to
prevent interleaving.  This means that libbase logging is safe to use
after fork().

Bug: 65062446
Bug: 153824050
Test: libbase/liblog unit tests
Test: new unit test for this specific scenario
Change-Id: I68014d7eff984dc2e5929a48a155931b843de9b3