Update patch set 2
Patch Set 2:

> how much difference does this make to the percentage?

Not much.

> i feel like -- and correct me if i'm wrong -- unlike the other changes in this stack, this is one that you wouldn't otherwise make. the others are all improvements, where as this feels like it makes the code a bit worse to work around the tool limitation?

I think I'd still make most of these changes: the smartsocket stuff is particularly confusing, since it appears that it's used in adbd because of the `#if !ADB_HOST` it contains, but it doesn't even make any sense to be used on the device.

Patch-set: 2
Reviewer: Gerrit User 1079148 <1079148@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0