Ignore top 16 bits in vaddr for the TBI enabled device
This diff mask lower 48bits on AArch64 device to avoid noise from top
byte.
ARM's Top-byte Ignore feature is available in all Armv8 AArch64 hardware.
This feature means that the hardware ignores the top byte of a pointer
when accessing memory. The physical address space was also defined to
support up to 48 bits.

Change-Id: Id639f4385f5dcf5e62417b75996d3146f782fb0a