liblog: remove locks for logd_socket
logd_socket historically used atomics to manage its lifetime.  They
were unfortunately unsafe and later replace with a RwLock.  A lock is
also problematic as it is too heavy weight for the typical use case
and implies that logging is neither async nor fork safe.

This change uses the fact that DGRAM sockets can have connect() called
on them multiple times.  It therefore uses a single socket for the
duration of the problem.  This socket is thread/async/fork safely
created when the first log message is sent.

The one caveat is __android_log_close(), which is intended only to be
used by zygote when it is single threaded and is therefore not thread
safe.  It will close this socket and reset the underlying variable,
such that the next log message will go through the above
initialization.

Bug: TBD
Test: Logging works
Test: unit test TBD
Change-Id: Ia4dbf7479dbe50683d124558ab2f83bff53b8f5f