Fix missing libsnapshot_test libraries
VTS tests have both the 32 and 64 bit architectures supported, but it
may happen that the system image does not provide both the libraries
required by the tests.
This patch fixes the failing linker when loading vts_libsnapshot_test
due to the lack of:
- android.hardware.boog@{1.0,1.1}
- libsparse
by importing them as static libraries instead of dynamic.

Test: manual inspection with ldd, VTS on GSI-via-DSU
Bug: 148889015
Signed-off-by: Alessio Balsini <balsini@google.com>
Change-Id: I0c5df2e981181ee9a5bc9fd0a6a610b143595ae6