Update patch set 1
Patch Set 1:

(1 comment)

We suppress ENOENT from `mount` but not ENODEV which it looks like is the errno that we'd get if `mount` fails, so I'm guessing that this does generate an error.

We have more than a few errors from init builtins depending on the configuration of the device already though.  If this only causes errors on older kernels, I think it will be okay.  It's likely better to see a few errors now than deal with the duplication to try to avoid the error.

Patch-set: 1
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>