logd: refactor mLast setting into a GetOldest function
This code was duplicated throughout LogBuffer.cpp, so refactor it into
a single location and clean up some comments along the way.

Pending verification: this may fix a subtle bug: if `logcat -c` is
used from a non-privileged UID, the current code may set mLast to the
oldest seen log message *from that UID* and not not the oldest log
message for that log id.  That may prevent pruning from the start of
that log.

Test: logging works
Change-Id: I1749293ce6ea1697dd8a9258cfd7eab29dbeac6e