Update patch set 2
Patch Set 2: Presubmit-Ready+1

Strange that it cleared the +2/+PV, I only updated the commit message.

I was indeed able to reproduce this issue.  It's a potential root cause for b/144382260 as well, so I added that bug to the CL.

Patch-set: 2
Reviewer: Gerrit User 1064128 <1064128@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=+1