fastbootd: move AVB footer to end of block device
If the flashed image is smaller than the block device, the AVB footer
will not be at the end of the partition. Although images are normally
created to match the partition size the GKI boot.img must work on all
devices, and the size of the boot partition will vary.

Move the AVB footer to the end of the partition before flashing, if it
is not there already.

Bug: 156036850
Change-Id: I11f0c7d32d1b6c74edd4f84f815d175605280cb8