libutils_singleton_test: Remove nonexistent required modules
The ":<module name>" syntax is for the "data:" property.
"required:" does not recognize the ":" syntax, so the required
entries of libutils_singleton_test doesn't have any effect:
* ":libutils_test_singleton1"
* ":libutils_test_singleton2"

Removing the leading ":" doesn't fix this because the test is
dlopen-ing the libraries from the test's data/ storage.
Changing "required:" to "data:" still wouldn't work because it
raises a build break due to missing variant / mutator.
A possible fix is described in b/128018586 where we modify the
build rule to allow cc_test_library be package into test suites

Bug: 7456955
Bug: 128018586
Test: atest libutils_singleton_test
Change-Id: I50d588e73056298e36e40c6a89af85db3bb71912