libsparse: Limit merged block size
On the operating system doesn't support 'mmap64' (e.g.: WIN32), the
malloc is used to replace the mmap feature in the write_fd_chunk
implementation. However, when the merged chunk size grows to several
gigabytes, which usually causes malloc returns out of memory failure.
To fix this, limit the merged block size to 64 MB to avoid generating
chunks in tremendous size.

Bug: 156057250
Change-Id: Ic4edb963a3d99f718d7630aba3f351729a84e994