Update patch set 4
Patch Set 4:

> Patch Set 4: Code-Review+2
> 
> Note that boot images are usually flashed in the bootloader, not fastbootd (flashstation and host fastboot only issue "reboot fastboot" before flashing dynamic partitions).
> 
> If that's going to be a problem, it might be better to do this logic in the host tool.

This is just for compliance testing with GKI in R for now so this can be covered in the flashing instructions to QA teams. Doing it in the host tool sounds ideal though I wonder whether it knows the size of the device's boot partition. Will investigate as time permits.

Patch-set: 4
Reviewer: Gerrit User 1131915 <1131915@85c56323-6fa9-3386-8a01-6480fb634889>
Label: Presubmit-Ready=0