Update patch set 13
Patch Set 13:

It looks like the performance of this is better on the current benchmark, but it's a bit worse when running on the same map or just a few maps different.

I've uploaded this cl:

https://android-review.googlesource.com/c/platform/system/core/+/1314538

With a bunch of new benchmarks for this.

The most likely case is only a few new maps, and the performance change was noticeable enough that this probably needs to be addressed.

Patch-set: 13
Reviewer: Gerrit User 1019050 <1019050@85c56323-6fa9-3386-8a01-6480fb634889>