Update patch set 1
Patch Set 1:

Thanks for the +2 Josh!

I was going to create a buganizer ticket for this and attach the trace with persist.adb.trace_mask=1 also. I'll add you to that and revise this commit text with the Bug: tag. Basically I performed the test as follows:

svc usb setFunctions mtp  <-- sets USB ConfigFS to "mtp,adb" mode
rm /config/usb_gadget/g1/configs/b.1/f1  <-- forces a FUNCTIONFS_UNBIND

Trace clearly shows there was a write_request after the offline event which is where it gets stuck.

> I suspect there's still a race here, though, since the error callback can get run on the worker thread as a result of a read failing, while the main thread is about to submit a write. We might have to just stop doing opportunistic writes entirely.

Ah quite possible. Any idea if this is even possible to stimulate in a test?

> (Is there any chance we can make io_submit fail instead of block when the endpoint is dead? I can't think of a situation where this behavior could actually be useful for anyone...)

That was my plan B. I remember mentioning this to you a couple years ago, and at the time I did end up implementing this in our downstream f_fs.c kernel driver but I never got around to polishing it to send to upstream or ACK, and it still feels kinda hacky to me. On our current development target we haven't propagated this fix, but the original issue of frequent offline stalls was mitigated with your refactoring to use a worker thread for the bulk of the AIO read/writes, which is pthread_kill-able. So now this issue is coming only when the main thread itself is performing a write to FFS. Hard to say if this is a general enough problem to fix once and for all in kernel but I can give it a shot as well.

Patch-set: 1