Modify mount_all to set ro.fstab
Currently the ReadDefaultFstab function, which calls GetDefaultFstab,
makes some assumptions about what the fstab will be called and where
it is located. This is being used by vold to set up userdata encryption
and for gsid, and is even used in the default boot control HAL, so it
has become quite baked.

The original way for a board to specify things to mount was to use the
"mount_all /path/to/fstab" command in init.rc. However, due to the
above functionality, the path after mount_all is no longer very useful,
as it cannot differ from the inferred path, or userdata encryption and
other features will be broken.

On Cuttlefish, we have an interest in being able to test alternative
userdata configurations (ext4 vs f2fs, encryption on/off, etc.) and
currently the only way to achieve this is to either a) modify the
ro.hardware or ro.platform.hardware properties, which breaks a bunch
of things like default HAL filenames, or regenerate our odm.img or
vendor.img filesystems. We can't simply install another fstab and
point to it with "mount_all".

This change fixes the default behavior by setting up the ro.fstab
property after the last "mount_all" command is issued, using the path
specified by the init.rc. Nearly every board uses mount_all, and as
mentioned above, all of them use one of the GetDefaultFstab implicit
paths (otherwise boot will be broken.) So, really, using the ro.fstab
in GetDefaultFstab is always going to work, and we don't need to check
the other properties any more.

If mount_all isn't called, nothing changes, and the fallback properties
will be used instead.

Bug: 142424832
Change-Id: I3e10f66aecfcd48bdb9ebf1d304b7aae745cbd3c