Update patch set 1
Patch Set 1:

> Patch Set 1:
> 
> I understand the issue, but I don't think we can solve it this way.  Pixel phones call mount_all multiple times with different fstabs to set up different types of partitions, so ro.fstab would be ambiguous for them.

I'm not sure this really matters. The other fstab files it uses are fragments, and only one of them can be the fstab needed by vold. Requiring this is set last seems to be OK; alternatively, we could only update the property on --late? I think given the 'trigger' semantics of "--late" (or no argument), it would not be useful to call this more than once anyway.

> As an alternative: what if we keep reading "ro.fstab" in ReadDefaultFstab() and cuttlefish calls `mount_all ${ro.fstab}` instead of hardcoding the path.  Then you'd be able to change the fstab for both ReadDefaultFstab() and mounting via that property, but we won't have the ambiguity on pixel.

I think the way it works in all the pixel init.rc files it would work fine with my change (they all call "--late" last, and that's always the right fstab).

Looking at the alternative, we already have a boot property we remap to a call to mount_all, so certainly, that could work.

Android originally had no ReadDefaultFstab(), and mount_all has been around forever. Really the way to fix this properly is to either do it the way I'm suggesting here, or make the "mount_all" path optional, so it can use ReadDefaultFstab() if it isn't provided.

Otherwise, I think it is pretty fragile that we allow arbitrary paths to be passed to "mount_all", but at least one of those paths must match one of the magic property names..

Patch-set: 1